+++ Abc.HabitTracker.Api/Repository/ILogsRepository.cs
@@ -7,9 +7,9 @@ namespace Abc.HabitTracker.Api.Repository
     public interface ILogsRepository
     {
         Logs CreateLogs(Logs logs);
-        Int16 GetCurrentStreak(Guid HabitId);
-        Int16 GetLongestStreak(Guid HabitId);
-        Int16 GetLogCount(Guid HabitId);
+        Int32 GetCurrentStreak(Guid HabitId);
+        Int32 GetLongestStreak(Guid HabitId);
+        Int32 GetLogCount(Guid HabitId);
         List<DateTime> GetAllLogsTime(Guid HabitId);
 
         Dictionary<Guid, List<DateTime>> GetHabitAndLogsFromUserID(Guid UserID);
