+++ Abc.HabitTracker.Api/Service/ILogsService.cs
@@ -9,11 +9,11 @@ namespace Abc.HabitTracker.Api.Service
     {
         Logs CreateLogs(Logs logs);
 
-        Int16 GetCurrentStreak(Guid HabitId);
+        Int32 GetCurrentStreak(Guid HabitId);
 
-        Int16 GetLongestStreak(Guid HabitId);
+        Int32 GetLongestStreak(Guid HabitId);
 
-        Int16 GetLogCount(Guid HabitId);
+        Int32 GetLogCount(Guid HabitId);
 
         List<DateTime> GetAllLogsTime(Guid HabitId);
